pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "deepscientist"
version = "0.1.0"
description = "DeepScientist Core skeleton"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "mcp>=1.19,<2",
  "Pillow>=10,<12",
  "PyYAML>=6,<7"
]

[project.scripts]
ds = "deepscientist.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
